+++ CumulusMX/Stations/JsonStation.cs
@@ -801,10 +801,10 @@ namespace CumulusMX.Stations
 							if (cumulus.LaserDepthBaseline[rec.index] == -1)
 							{
 								// MX is not calculating depth
-
 								decimal? depth = rec.depth.HasValue ? ConvertUnits.LaserMmToUser(rec.depth.Value * multiplier) : null;
 								station.DoLaserDepth(depth, rec.index);
 							}
+							// else DoLaserDistance() calcs the depth
 						}
 						catch (Exception ex)
 						{